+++ src/__tests__/constants.test.ts
@@ -1,4 +1,5 @@
-import { getControlTreatmentsWithConfig, CONTROL_WITH_CONFIG } from '../constants';
+import { CONTROL_WITH_CONFIG } from '../constants';
+import { getControlTreatmentsWithConfig } from '../utils';
 import SplitIO from '@splitsoftware/splitio/types/splitio';
 
 describe('getControlTreatmentsWithConfig', () => {
